rename database dir to postgresDb
