+++ postgresDb/credentials/credentials.js
@@ -1,0 +1,6 @@
+const credentials = {
+  username: 'alexsimrell',
+  password: '',
+};
+
+module.exports = credentials;
